Add files via upload
The goal of this project is to build, train, and evaluate a Convolutional Neural Network (CNN) model to classify handwritten digits from the MNIST dataset. The MNIST dataset is a widely used benchmark in the field of machine learning and computer vision, consisting of 28x28 grayscale images of handwritten digits (0-9). The project also includes visualizing the model's predictions to gain insights into its performance.